rework ping
